+++ src/rules.ts
@@ -20,6 +20,7 @@ import './html/undefined-elements';
 
 import './polymer/behaviors-spelling';
 import './polymer/call-super-in-callbacks';
+import './polymer/content-to-slot';
 import './polymer/databinding-calls-must-be-functions';
 import './polymer/databind-with-unknown-property';
 import './polymer/element-before-dom-module';